Added docstrings in DatabaseManager class

+++ python-backend/src/DineFinderAI/db/DatabaseManager.py
@@ -9,20 +9,54 @@ import pandas as pd
 
 
 class DatabaseManager:
+  """
+  A class to manage database operations and optionally interact with a JSON data file.
+
+  Attributes:
+      database_filepath (Path): The file path to the database file.
+      json_filepath (Path | None): The file path to a JSON data file (optional).
+  """
   def __init__(self, database_filepath: Path, json_filepath: Path | None = None) -> None:
+    """
+    Initializes the DatabaseManager with the paths to the database file and an optional JSON data file.
+
+    Args:
+        database_filepath (Path): The file path to the database file.
+        json_filepath (Path | None, optional): The file path to a JSON data file. Defaults to None.
+    """
     self.database_filepath = database_filepath
     self.json_filepath = json_filepath
 
   def connectFunc(self) -> None:
+    """ Establishes a connection to the database. """
     self.con = sqlite3.connect(self.database_filepath)
 
   def closeFunc(self) -> None:
+    """ Closes the database connection. """
     self.con.close()
 
   def execute(self, query: str) -> pd.DataFrame:
+    """
+    Executes an SQL query and returns the result as a pandas DataFrame.
+
+    Args:
+        query (str): The SQL query string to execute on the database.
+
+    Returns:
+        pd.DataFrame: A DataFrame containing the results of the SQL query.
+    """
     return pd.read_sql_query(query, self.con)
 
   def insertData(self) -> None:
+    """
+    Reads Restaurant data from a JSON file, filters the entries and stores the entries into the database.
+
+    The following filters are applied:
+        - Only businesses with the "Restaurants" category are kept.
+        - Businesses with certain categories (e.g., "Beauty & Spas", "Health & Medical", etc.) are excluded.
+        - Businesses with missing or empty "name" or "address" fields are excluded.
+        - Businesses with "Wellness" in their name are excluded.
+    """
     if not self.json_filepath:
       print("Please provide a json file with data before trying to insert data...")
       return
@@ -66,17 +100,42 @@ class DatabaseManager:
     return df
 
   def addColumn(self, column_name: str, sql_data_type: str) -> None:
+    """
+    Adds a new column to the 'restaurants' table in the database.
+
+    Args:
+        column_name (str): The name of the column to add.
+        sql_data_type (str): The SQL data type of the new column (e.g., 'TEXT', 'INTEGER').
+    """
     query = f"ALTER TABLE restaurants ADD {column_name} {sql_data_type}"
     self.con.cursor().execute(query)
 
-def main() -> None:
+def confirmDeletion() -> bool:
+  """
+  Prompts the user to confirm whether they want to delete the existing database file.
 
+  Returns:
+      bool: Returns `True` if the user confirms deletion by entering 'y', and `False` if 
+            the user declines by entering 'n'.
+  """
+  confirm: str = input("Do you want to delete the existing database file? [y|n] -> ")
+  while confirm != "y" and confirm != "n":
+    confirm = input("Please type either 'y' or 'n' -> ")
+
+  return confirm == "y"
+
+def main() -> None:
   resources_path = Path.cwd().joinpath("resources")
   db_path = resources_path.joinpath("database.db")
   data_path = resources_path.joinpath("yelp_academic_dataset_business.json")
   if not data_path.exists():
     print("Data path doesn't exist, please provide one.")
     return
+  
+  if db_path.exists() and not confirmDeletion():
+    print("Abort")
+    return
+
   db_path.unlink(missing_ok=True)
 
   db_manager = DatabaseManager(db_path, data_path)
